refactor: 从 func/index 中拆分一部分代码出去
