+++ src/func/index.ts
@@ -7,6 +7,7 @@ import { info, warn, error, i18n, lute } from "../utils";
 import * as serverApi from '../serverApi';
 import { reservation, settings } from '../global-status';
 import { Retrieve, RetvFactory } from './reserve';
+import { getDailynoteSprig, renderDailynotePath } from './dailynote';
 
 
 const default_sprig = `/daily note/{{now | date "2006/01"}}/{{now | date "2006-01-02"}}`
@@ -103,26 +104,7 @@ export async function currentDiaryStatus() {
 }
 
 
-/**
- * 
- * @param notebook 笔记本对象
- * @returns 笔记本的 Daily Note 路径模板
- */
-export async function getDailynoteSprig(notebookId: string): Promise<string> {
-    let conf = await serverApi.getNotebookConf(notebookId);
-    let sprig: string = conf.conf.dailyNoteSavePath;
-    return sprig;
-}
 
-
-/**
- * 要求思源解析模板
- * @param sprig
- * @returns 
- */
-export async function renderDailynotePath(sprig: string) {
-    return await serverApi.renderSprig(sprig);
-}
 
 /**
  * getDocsByHpath returns all documents in the database that have a given hpath.
